+++ src/rust/Cargo.toml
@@ -33,7 +33,7 @@ members = [
 ]
 
 [profile.dev]
-opt-level = 0
+opt-level = 2
 lto       = false
 debug     = true
